refactor: split dce state and methods
