+++ src/optimize/control/dce/methods.rs
@@ -1,0 +1,19 @@
+use super::*;
+
+pub(crate) fn dce_method(method: ClassMethod, class_name: &str, parent_name: Option<&str>) -> ClassMethod {
+    let context = ClassEffectContext {
+        class_name: class_name.to_string(),
+        parent_name: parent_name.map(str::to_string),
+    };
+    ClassMethod {
+        body: with_class_effect_context(Some(context), || dce_block(method.body)),
+        ..method
+    }
+}
+
+pub(crate) fn dce_method_without_context(method: ClassMethod) -> ClassMethod {
+    ClassMethod {
+        body: with_class_effect_context(None, || dce_block(method.body)),
+        ..method
+    }
+}
